backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
redis
rq
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart